requirements.txt
aiohttp
aiosignal
asgiref
async-timeout
asyncio
attrs
certifi
cffi
charset-normalizer
colorama
cryptography
dj-rest-auth
Django
django-cors-headers
django-filter
django-rest-knox
djangorestframework
djangorestframework-simplejwt
frozenlist
gunicorn
idna
Markdown
multidict==4.0
Pillow
pycparser
PyJWT
pytz
requests
sqlparse
tqdm
tzdata
urllib3
whitenoise
yarl
django-dotenv